microtouch3m-cli: also measure noise using peak-to-stray
Peak-to-peak noise makes sense when we don't apply any correction to
the signal.

If we do apply stray correction to the signal, we already get an ideal
'0' noise level to which we compare the signal we receive. We can
therefore measure the amount of noise above the stray, which is what
may end up affecting the logic that detects touches (as the signal
will always increase when a touch is performed).