microtouch3m-scope: don't error out if sensitivity level fails
Each 'sensitivity level' (from 0 to 6) maps a specific combination of
extended sensitivity settings. If the user sets a different set of
extended sensitivity values making the new combination not match one
expected 'sensitivity level', the reading of the value will fail.

So, just go on if we get an error reading the 'sensitivity level' as
we're anyway going to show all the extended values anyway.